Slight reorganization
Separated the action io functions into the io module; this is is with an eye towards the browserview version
