+++ lib/actions.js
@@ -5,12 +5,8 @@
 /*
 *
 * TODO:
-* - is it necessary to reorganize by pushing the io functions into a separate module?
 * - add the new field to the Browserify version, and expand the 'bridge' part to include the action handling
-* - documentation
 */
-
-const fs = require('fs');
 
 /**
  * Class to maintain actions.
@@ -115,72 +111,5 @@ class ActionList {
     }
 }
 
-/**
- * Get the action list (if any). For the time being only file input is used.
- *
- * @param {object} conf - Overall configuration; the only field that matter here is "conf.actions"
- * @returns {Promise} - a promise containing the nicknames as an object parsed from JSON.
- */
-function get_action_list(conf) {
-    return new Promise((resolve, reject) => {
-        if (conf.actions) {
-            // This is a local file. Use an async function to retrieve the file,
-            // though a sync function would work just as well
-            // However if, at some point, remote access is added, then we have the right
-            // structure.
-            fs.readFile(conf.actions, 'utf-8', (err, body) => {
-                if (err) {
-                    reject(new Error(`problem accessing local file ${conf.actions}: ${err}`));
-                } else {
-                    // Try to parse the content as JSON and, if it works, that is almost
-                    // the final result, module turn all the nicknames to lowercase
-                    let json_content = {};
-                    try {
-                        json_content = JSON.parse(body);
-                    } catch (parse_err) {
-                        throw new Error(`JSON parsing error in ${conf.actions}: ${parse_err}`);
-                    }
-                    resolve(json_content);
-                }
-            });
-        } else {
-            resolve({});
-        }
-    });
-}
 
-/**
- * Store the action list (if any) in a file.
- *
- * @param {object} conf - Overall configuration; the only field that matter here is "conf.actions"
- * @param {ActionList} actions - the data to be stored
- * @returns {Promise} - a promise returning no value per se
- */
-function store_action_list(conf, actions) {
-    return new Promise((resolve, reject) => {
-        if (conf.actions) {
-            // This is a local file. Use an async function to retrieve the file,
-            // though, I believe, a sync function would work just as well
-            // However if, at some point, remote access is added, then we have the right
-            // structure.
-            // The action of copying is done via a sync function, though. It is simpler...
-            try {
-                fs.copyFileSync(conf.actions, `${conf.actions}.bak`);
-            } catch (err) {
-                reject(new Error(`problem copying local file ${conf.actions} for backup: ${err}`));
-            }
-            fs.writeFile(conf.actions, actions.toStore(), 'utf-8', (err) => {
-                if (err) {
-                    reject(new Error(`problem writing local file ${conf.actions}: ${err}`));
-                } else {
-                    resolve();
-                }
-            });
-        } else {
-            resolve();
-        }
-    });
-}
-
-
-module.exports = { ActionList, get_action_list, store_action_list };
+module.exports = { ActionList };
