+++ lib/io.js
@@ -20,56 +20,10 @@ const _        = require('underscore');
 const Octokat  = require('octokat');
 const validUrl = require('valid-url');
 
-/**
- * Get the IRC log. The input provided in the configuration is examined whether
- * it is a URL (in which case this is retrived via HTTP) or not (in which case
- * it is considered to be a local file). Returns a Promise with the content
- * of the input as one string.
- *
- * @param {object} conf - Overall configuration; the only field that matter
- *     here is "conf.input"
- * @returns {Promise} - a promise containing the irc log as a single string.
- */
-exports.get_irc_log = (conf) => new Promise((resolve, reject) => {
-    if (conf.irclog) {
-        // This field may be present if called from a CGI script, and the IRC log
-        // is uploaded
-        resolve(conf.irclog);
-    } else if (url.parse(conf.input).protocol !== null) {
-        // This is a resource on the Web that must be fetched
-        fetch(conf.input)
-            .then((response) => {
-                if (response.ok) {
-                    if (response.headers.get('content-type') === 'text/plain') {
-                        return response.text();
-                    }
-                    reject(new Error('IRC log must be of type text/plain'));
-                } else {
-                    reject(new Error(`HTTP response ${response.status}: ${response.statusText}`));
-                }
-            })
-            .then((body) => {
-                // Resolve the returned Promise
-                resolve(body);
-            })
-            .catch((err) => {
-                reject(new Error(`problem accessing remote file ${conf.input}: ${err.message}`));
-            });
-    } else {
-        // This is a local file. Use an async function to retrieve the file,
-        // though, I believe, a sync function would work just as well
-        fs.readFile(conf.input, 'utf-8', (err, body) => {
-            if (err) {
-                reject(new Error(`problem access local file ${conf.input}: ${err}`));
-            } else {
-                resolve(body);
-            }
-        });
-    }
-});
 
+/* ------------------------------------ General Utilities ------------------------- */
 /**
-* Basic sanity check on the URL.
+* Basic sanity check on a URL.
 *
 * The function returns a (possibly slightly modified) version of the URL if
 * everything is fine, or a null value if the input argument is not a URL (but
@@ -124,6 +78,149 @@ function check_url(address) {
     // If we got this far, this is a proper URL, ready to be used.
     return retval;
 }
+
+
+/**
+ * Committing new data on the github repo
+ *
+ * The following terms in the configuration are relevant for this function
+ * - ghrepo: the full name of the repository. E.g., "w3c/scribejs"
+ * - ghpath: the path within the repository where the data must be stored. E.g., "test/minutes"
+ * - ghfname: the file name
+ * - ghname: the user name to be used when committing
+ * - ghemail: the user email to be used when committing
+ * - ghtoken: the user's OAUTH personal access token provided by GitHub
+ *             (see https://github.com/settings/tokens/new)
+ * - ghbranch: the target branch within the repository. This term may be
+ *             missing from the configuration, in which case the repo's default
+ *             branch is used
+ *
+ * @param {string} data - the markdown data to be uploaded
+ * @param {object} conf - the configuration containing additional data
+ * @returns {Promise} - the returned promise data is whatever the github API
+ *     returns; usable for debug
+ */
+// The GH access is done via the octokat library (https://www.npmjs.com/package/octokat). See
+// https://github.com/philschatz/octokat.js/tree/c6f38e5bd7b2dadeb9066eaae2e8a762354795d8/examples
+// for a set of examples which is an attempt to replace a proper documentation...
+//
+// This library provides a more concise layer on top of the official Github API:
+// https://developer.github.com/v3
+function commit(data, conf) {
+    // The token below is essential, it gives the access right to the repository.
+    // See GH documentation how to obtain it...
+    const gh               = new Octokat({ token: conf.ghtoken });
+    const repo             = conf.ghrepo.split('/');
+    const path             = `${conf.ghpath}/${conf.ghfname}`;
+    const path_with_branch = path + (conf.ghbranch ? `?ref=${conf.ghbranch}` : '');
+    // Before uploading we have to see if the file already exists or not. If it does
+    // we need the sha number of that resource; this is needed to be able to modify the
+    // content.
+    //
+    // This means: first we make a GET on the target resource; if the resources is there
+    // we retrieve the sha value from the return value, which must be added to the "message"
+    // object used to update or upload the real content. This last step is done
+    // in the 'upload' function.
+    return new Promise((resolve, reject) => {
+        // This function uploads the data with or without an sha value
+        const upload = (upload_resolve, upload_reject, sha = null) => {
+            const params = {
+                sha,
+                branch    : conf.ghbranch,
+                message   : conf.ghmessage,
+                content   : Buffer.from(data).toString('base64'),
+                committer : {
+                    name  : conf.ghname,
+                    email : conf.ghemail
+                }
+            };
+            // The gh...content idiom returns an object, and the 'add' verb is
+            // used to add content to it.
+            // The parameter is the full message for the GH API, including the
+            // real content (encoded)
+            gh.repos(...repo).contents(path).add(params)
+                .then((info) => {
+                    if (info && info.content && info.content.htmlUrl) {
+                        upload_resolve(info.content.htmlUrl);
+                    } else {
+                        upload_reject(info);
+                    }
+                })
+                .catch((err) => upload_reject(err));
+        };
+
+        // The gh...content idiom with a path returns an object that can be
+        // 'fetched' information about the content
+        gh.repos(...repo).contents(path_with_branch).fetch()
+            .then((info) => {
+                if (info) {
+                    // If the file really exists, then it has an sha that should be reused...
+                    upload(resolve, reject, info.sha ? info.sha : null);
+                }
+            })
+            .catch(() => {
+                // This means that file is unknown in the repository, ie,
+                // should be created from scratch
+                upload(resolve, reject);
+            });
+    });
+}
+
+/* ------------------------------------------------------------------------------------- */
+/* --------------------------------- Get the IRC log ----------------------------------- */
+/* ------------------------------------------------------------------------------------- */
+
+/**
+ * Get the IRC log. The input provided in the configuration is examined whether
+ * it is a URL (in which case this is retrieved via HTTP) or not (in which case
+ * it is considered to be a local file). Returns a Promise with the content
+ * of the input as one string.
+ *
+ * @param {object} conf - Overall configuration; the only field that matter
+ *     here is "conf.input"
+ * @returns {Promise} - a promise containing the irc log as a single string.
+ */
+exports.get_irc_log = (conf) => new Promise((resolve, reject) => {
+    if (conf.irclog) {
+        // This field may be present if called from a CGI script, and the IRC log
+        // is uploaded
+        resolve(conf.irclog);
+    } else if (url.parse(conf.input).protocol !== null) {
+        // This is a resource on the Web that must be fetched
+        fetch(conf.input)
+            .then((response) => {
+                if (response.ok) {
+                    if (response.headers.get('content-type') === 'text/plain') {
+                        return response.text();
+                    }
+                    reject(new Error('IRC log must be of type text/plain'));
+                } else {
+                    reject(new Error(`HTTP response ${response.status}: ${response.statusText}`));
+                }
+            })
+            .then((body) => {
+                // Resolve the returned Promise
+                resolve(body);
+            })
+            .catch((err) => {
+                reject(new Error(`problem accessing remote file ${conf.input}: ${err.message}`));
+            });
+    } else {
+        // This is a local file. Use an async function to retrieve the file,
+        // though, I believe, a sync function would work just as well
+        fs.readFile(conf.input, 'utf-8', (err, body) => {
+            if (err) {
+                reject(new Error(`problem access local file ${conf.input}: ${err}`));
+            } else {
+                resolve(body);
+            }
+        });
+    }
+});
+
+/* ------------------------------------------------------------------------------------- */
+/* --------------------------------- Get the nickname file ----------------------------- */
+/* ------------------------------------------------------------------------------------- */
 
 /**
  * Get the nickname mapping file (if any). The input provided in the
@@ -202,92 +299,9 @@ exports.get_nick_mapping = (conf) => {
     });
 };
 
-/**
- * Committing new data on the github repo
- *
- * The following terms in the configuration are relevant for this function
- * - ghrepo: the full name of the repository. E.g., "w3c/scribejs"
- * - ghpath: the path within the repository where the data must be stored. E.g., "test/minutes"
- * - ghfname: the file name
- * - ghname: the user name to be used when committing
- * - ghemail: the user email to be used when committing
- * - ghtoken: the user's OAUTH personal access token provided by GitHub
- *             (see https://github.com/settings/tokens/new)
- * - ghbranch: the target branch within the repository. This term may be
- *             missing from the configuration, in which case the repo's default
- *             branch is used
- *
- * @param {string} data - the markdown data to be uploaded
- * @param {object} conf - the configuration containing additional data
- * @returns {Promise} - the returned promise data is whatever the github API
- *     returns; usable for debug
- */
-// The GH access is done via the octokat library (https://www.npmjs.com/package/octokat). See
-// https://github.com/philschatz/octokat.js/tree/c6f38e5bd7b2dadeb9066eaae2e8a762354795d8/examples
-// for a set of examples which is an attempt to replace a proper documentation...
-//
-// This library provides a more consise layer on top of the official Github API:
-// https://developer.github.com/v3
-function commit(data, conf) {
-    // The token below is essential, it gives the access right to the repository.
-    // See GH documentation how to obtain it...
-    const gh               = new Octokat({ token: conf.ghtoken });
-    const repo             = conf.ghrepo.split('/');
-    const path             = `${conf.ghpath}/${conf.ghfname}`;
-    const path_with_branch = path + (conf.ghbranch ? `?ref=${conf.ghbranch}` : '');
-    // Before uploading we have to see if the file already exists or not. If it does
-    // we need the sha number of that resource; this is needed to be able to modify the
-    // content.
-    //
-    // This means: first we make a GET on the target resource; if the resources is there
-    // we retrieve the sha value from the return value, which must be added to the "message"
-    // object used to update or upload the real content. This last step is done
-    // in the 'upload' function.
-    return new Promise((resolve, reject) => {
-        // This function uploads the data with or without an sha value
-        const upload = (upload_resolve, upload_reject, sha = null) => {
-            const params = {
-                sha,
-                branch    : conf.ghbranch,
-                message   : conf.ghmessage,
-                content   : Buffer.from(data).toString('base64'),
-                committer : {
-                    name  : conf.ghname,
-                    email : conf.ghemail
-                }
-            };
-            // The gh...content idiom returns an object, and the 'add' verb is
-            // used to add content to it.
-            // The parameter is the full message for the GH API, including the
-            // real content (encoded)
-            gh.repos(...repo).contents(path).add(params)
-                .then((info) => {
-                    if (info && info.content && info.content.htmlUrl) {
-                        upload_resolve(info.content.htmlUrl);
-                    } else {
-                        upload_reject(info);
-                    }
-                })
-                .catch((err) => upload_reject(err));
-        };
-
-        // The gh...content idiom with a path returns an object that can be
-        // 'fetched' information about the content
-        gh.repos(...repo).contents(path_with_branch).fetch()
-            .then((info) => {
-                if (info) {
-                    // If the file really exists, then it has an sha that should be reused...
-                    upload(resolve, reject, info.sha ? info.sha : null);
-                }
-            })
-            .catch(() => {
-                // This means that file is unknown in the repository, ie,
-                // should be created from scratch
-                upload(resolve, reject);
-            });
-    });
-}
-
+/* ------------------------------------------------------------------------------------- */
+/* ------------------------------------ Output the minutes ----------------------------- */
+/* ------------------------------------------------------------------------------------- */
 /**
  * Output the minutes. Depending on the configuration, the values are stored in a file or on
  * a GitHub repository.
@@ -319,3 +333,72 @@ exports.output_minutes = (minutes, conf) => new Promise((resolve, reject) => {
         resolve('');
     }
 });
+
+/* ------------------------------------------------------------------------------------- */
+/* --------------------------------- Action files I/O ---------------------------------- */
+/* ------------------------------------------------------------------------------------- */
+
+
+/**
+ * Get the action list (if any). For the time being only file input is used.
+ *
+ * @param {object} conf - Overall configuration; the only field that matter here is "conf.actions"
+ * @returns {Promise} - a promise containing the nicknames as an object parsed from JSON.
+ */
+exports.get_action_list = (conf) => new Promise((resolve, reject) => {
+    if (conf.actions) {
+        // This is a local file. Use an async function to retrieve the file,
+        // though a sync function would work just as well
+        // However if, at some point, remote access is added, then we have the right
+        // structure.
+        fs.readFile(conf.actions, 'utf-8', (err, body) => {
+            if (err) {
+                reject(new Error(`problem accessing local file ${conf.actions}: ${err}`));
+            } else {
+                // Try to parse the content as JSON and, if it works, that is almost
+                // the final result, module turn all the nicknames to lowercase
+                let json_content = {};
+                try {
+                    json_content = JSON.parse(body);
+                } catch (parse_err) {
+                    throw new Error(`JSON parsing error in ${conf.actions}: ${parse_err}`);
+                }
+                resolve(json_content);
+            }
+        });
+    } else {
+        resolve({});
+    }
+});
+
+/**
+ * Store the action list (if any) in a file.
+ *
+ * @param {object} conf - Overall configuration; the only field that matter here is "conf.actions"
+ * @param {ActionList} actions - the data to be stored
+ * @returns {Promise} - a promise returning no value per se
+ */
+exports.store_action_list = (conf, actions) => new Promise((resolve, reject) => {
+    if (conf.actions) {
+        // This is a local file. Use an async function to retrieve the file,
+        // though, I believe, a sync function would work just as well
+        // However if, at some point, remote access is added, then we have the right
+        // structure.
+        // The action of copying is done via a sync function, though. It is simpler...
+        try {
+            fs.copyFileSync(conf.actions, `${conf.actions}.bak`);
+        } catch (err) {
+            reject(new Error(`problem copying local file ${conf.actions} for backup: ${err}`));
+        }
+        fs.writeFile(conf.actions, actions.toStore(), 'utf-8', (err) => {
+            if (err) {
+                reject(new Error(`problem writing local file ${conf.actions}: ${err}`));
+            } else {
+                resolve();
+            }
+        });
+    } else {
+        resolve();
+    }
+});
+
